+++ src/extension.ts
@@ -88,6 +88,7 @@ export async function activate (ctx: ExtensionContext) {
     commands.registerCommand('FastSfdc.retrieveSelected', cmds.retrieveSelected),
     commands.registerCommand('FastSfdc.retrieveSelectedMeta', cmds.retrieveSelectedMeta),
     commands.registerCommand('FastSfdc.deploy', cmds.deploy),
+    commands.registerCommand('FastSfdc.cancelDeploy', cmds.cancelDeploy),
     commands.registerCommand('FastSfdc.validate', () => cmds.deploy(true)),
     commands.registerCommand('FastSfdc.retrieveSingle', cmds.retrieveSelected),
     commands.registerCommand('FastSfdc.configureStaticResourceBundles', cmds.configureStaticResourceBundles),